Fix compilation warnings for Win64 platform
Make explicit int and pointer type casts where necessary.
Add missing includes for type visibility.
Note: M_PI constant requires _USE_MATH_DEFINES macro defined (will be
added later).

+++ src/RenderStack.h
@@ -45,7 +45,7 @@ public:
     }
 
     static int depth() {
-        return stack.size();
+        return static_cast<int>(stack.size());
     }
 
 private:
